Refactor executables to separate crates
