+++ gas/src/main.rs
@@ -1,11 +1,11 @@
 extern crate parity_wasm;
 extern crate pwasm_utils as utils;
+extern crate logger;
 
 use std::env;
 
 fn main() {
-
-	utils::init_log();
+	logger::init_log();
 
 	let args = env::args().collect::<Vec<_>>();
 	if args.len() != 3 {
