--- src/logger.rs
@@ -1,26 +1,0 @@
-extern crate log;
-
-use std::env;
-use log::LogLevelFilter;
-use env_logger::LogBuilder;
-
-lazy_static! {
-	static ref LOG_DUMMY: bool = {
-		let mut builder = LogBuilder::new();
-		builder.filter(None, LogLevelFilter::Info);
-
-		if let Ok(log) = env::var("RUST_LOG") {
-			builder.parse(&log);
-		}
-
-		if let Ok(_) = builder.init() {
-			trace!("logger initialized");
-		}
-		true
-	};
-}
-
-/// Intialize log with default settings
-pub fn init_log() {
-	let _ = *LOG_DUMMY;
-}
